refactor: change chapter number and latest chapter number types from float to int for consistency feat: add MugiwaraOficialChecker service for manga chapter checking and information extraction

+++ back-end/MangaWhisper.Domain/Entities/Chapter.cs
@@ -4,7 +4,7 @@ public class Chapter
 {
     public int Id { get; set; }
     public int MangaId { get; set; }
-    public float Number { get; set; }
+    public int Number { get; set; }
     public string Title { get; set; } = string.Empty;
     public string Url { get; set; } = string.Empty;
